refactor(desktop): split browser backend headless modules
